+++ Symulacja Cyfrowa/event.cpp
@@ -1,7 +1,7 @@
 #include "event.h"
 
-Event::Event(float event_time, Network* network) : event_time_(event_time), network_(network) {}
+Event::Event(int event_time, Network* network) : event_time_(event_time), network_(network) {}
 
-float Event::get_time() {
+int Event::get_time() {
 	return event_time_;
 }
