+++ Symulacja Cyfrowa/user_end_of_service_event.h
@@ -3,7 +3,7 @@
 #include"event.h"
 class UserEndOfServiceEvent : public Event {
 public:
-	UserEndOfServiceEvent(float event_time, Network* network, Client* client_to_remove);
+	UserEndOfServiceEvent(int event_time, Network* network, Client* client_to_remove);
 	void Execute() override;
 private:
 	Client* user_to_remove_;
